+++ include/JAInter/Object.h
@@ -7,7 +7,7 @@
 namespace JAInter {
 struct Object : public ObjectBase {
 	virtual ~Object();                     // _00
-	virtual void startSound(ulong, ulong); // _04
+	virtual void startSound(u32, u32); // _04
 	virtual void disable();                // _1C
 	virtual void loop();                   // _30
